requirements.txt
pyqt5
pyrebase4
firebase-admin
requests
moviepy
